Make relative font calculations (smaller and larger) more precise. Smaller font for base value 3.75 (i.e. 3.75 / 1.2) did not result in 3.125
Autoported commit.
Original commit hash: [86f6334c]

+++ itext/itext.html2pdf/itext/html2pdf/css/apply/util/FontStyleApplierUtil.cs
@@ -252,11 +252,11 @@ namespace iText.Html2pdf.Css.Apply.Util {
 
         public static float ParseRelativeFontSize(String relativeFontSizeValue, float baseValue) {
             if (CssConstants.SMALLER.Equals(relativeFontSizeValue)) {
-                return baseValue / 1.2f;
+                return (float)(baseValue / 1.2);
             }
             else {
                 if (CssConstants.LARGER.Equals(relativeFontSizeValue)) {
-                    return baseValue * 1.2f;
+                    return (float)(baseValue * 1.2);
                 }
             }
             return CssUtils.ParseRelativeValue(relativeFontSizeValue, baseValue);
